Fix #76, add mpool implementation for BPv7
BPv7 structures will require a significantly different storage mechanism
from what BPv6 has used.  This is a proof-of-concept implementation to
provide storage via chaining together groups of 256-byte blocks, with
a preallocated pool and a fast allocator/deallocator.

This allows the user to choose the amount of RAM dedicated to bundle
storage and reduces dependency on the heap. Fixed size blocks should
eliminate any risk of fragmentation in the pool.

This also updates the CMake file to include hooks for building the v7
implementation (which will be extended as more functionality gets added).