Fix #91, initial implementation of job queues
Contains several enhancements to make work queues more abstract and
easier to manage.

This also enhances the bplib_mpool_block_t structure such that every
instance of this structure is traceable back to its parent block as
well as the pool it originally came from.  This alleviates the need
to pass the pool pointer separately on many APIs, simplifying things.

Introduces another block type for a "job" which is abstracted to
represent any schedulable item.  This was formerly only for flows,
but now can be used for any purpose.  State changes of interfaces
are now handled through this method.