Fix #136, do not set link libs on object lib
The cmake script was incorrectly setting the target_link_libraries
on the object library.  Although this code is what requires the
library, only the compile flags are needed here.  The link libs
are required at the parent library link, and they are already
there (so this is redundant).

New CMake versions appear to have simply ignored this line, but
old CMake versions produce an error.