Fix #146, skip call to cbor_parser_init on payloads
As part of decoding, the function `cbor_parser_init()` was always
invoked on the bundle payload, regardless of what type of block it was.
This parser was used for extension blocks that are CBOR encoded, or
discard if it was not a CBOR block (e.g. payload).

However, in the TinyCBOR library, this init function will preemptively
decode the first element.  If the block is not actually CBOR, this is
not defined.

Large file transfers worked OK until offset 0x1bffffff (about 469MB)
because tag byte values of 0x00 - 0x1b are valid CBOR.  However, at
offset 0x1C000000, this would fail, because byte 0x1C is not a valid
CBOR tag byte.  By skipping the cbor_parser_init call entirely, this
avoids the problem and is more correct as well.