Fix #158, do not update stream_pos on retry attempts
The "stream_pos" on the sender side should only be updated on the first
attempt to send a block of data via bplib_send().  On subsequent retries
it should only call bplib_send() again.

This also adds information to the debug messages to show the offsets and
positions so the activity in this area can be more clearly observed.