Fix #167, provide config/query API
Implements a basic integer config/query API that can be used for any
runtime or configuration variable.  Initially this is used to replace
calls to bplib_os memory functions that reported the memory usage.

This keeps the code layering correct and also adds a stub point for
testing other dependent software.