Fix #184, add build and test workflow
Workflow to build and test BPLib in a variety of configurations.  The
workflow implements a matrix of Debug and Release build types, and OSAL
and POSIX adapter layers.

For the OSAL builds, the unit tests are also enabled, and the tests are
executed.

Additionally, for the Debug OSAL build, the coverage statistics are
computed and the results are included in the summary report, so they can
be monitored.

This also migrates the "bptest" test program into a proper functional
test based on UT Assert.