Fix #216, split v7_codec.c into parts
This splits the monolithic "v7_codec.c" file, which contained
encode/decode routines for all BPv7 elements, into smaller components.

There is a set of fundamental elements, such as integers, crc, containers
and bitmaps, as well as a dedicated source unit for each block type and
admin record.  In the future the software may allow for dynamic
registration of additional block types and admin records.