Fix #229, correct cast alignment warnings
Most of these are fixable by adding an intermediate "void *" cast.

For the case of casting to the "bplib_mpool_block_content_t" type, which
occurs somewhat frequently in the mpool logic, use the cast function to do
this conversion so.  This reduces the number of times this complexity
occurs in the source.

Changing this exposed another issue with the check for
"bplib_mpool_is_any_content_node()", in that it did not include all the
block types for which this is valid - api was not included.  This is
also corrected here.