Fix #245, implement safer codec loops
Implement the encoding and decoding of CBOR containers as a loop.  This
permits the "error" flag to be easily polled after each iteration, and
if it is ever set, the operation should exit safely.

This is important because after any encode/decode issue, the tiny CBOR
state will no longer be in sync with the data, and this library may
assert if invoked with a bad state object.