Fix #252, propagate error code after decode failure
The "bplib_generic_bundle_ingress()" function was not correctly
propagating an error status from the underlying decode operation.  It
was logging the error but using the return value from the bplog macro,
which will be BP_SUCCESS.

This also updates the coverage tests to use "INT32_EQ" rather than
"UINT32_EQ", as the return value from these API calls is a signed
integer.